create a function to sort by bbs sorting criteria and seperate post.js to post.js & editpost.js
