+++ src/main/resources/application.properties
@@ -7,7 +7,7 @@ spring.datasource.url=jdbc:oracle:thin:@localhost:1521:orcl
 spring.datasource.username=test
 spring.datasource.password=test
 
-spring.jpa.hibernate.ddl-auto=validate
+spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
 #spring.mvc.converters.preferred-json-mapper=gson
