+++ src/main/resources/static/js/adminpage.js
