+++ src/main/resources/templates/admin/admin_member.html
@@ -30,7 +30,8 @@
             </div>
             <span class="d-flex" th:text="'총 회원 수 : ' + ${memberList.getTotalElements()}"></span>
         </div>
-        <table id="table" class="flex-column justify-content-around align-items-stretch table table-striped table-responsive table-hover">
+        <table id="table" class="flex-column justify-content-around align-items-stretch
+        table table-striped table-responsive table-hover">
             <thead>
             <tr>
                 <th style="width: 5%;">No.</th>
