+++ src/main/resources/templates/post/editBbs.html
@@ -51,7 +51,7 @@
 </div>
 <th:block layout:fragment="script">
     <script src="https://code.jquery.com/jquery-3.6.0.js"></script>
-    <script src="/js/post.js"></script>
+    <script src="/js/editpost.js"></script>
     <script type="text/javascript">
         window.onload = function () {
             $('#bbsContents').summernote(setting);
