+++ src/main/resources/templates/post/post.html
@@ -12,9 +12,21 @@
 <div layout:fragment="content">
     <div layout:replace="fragments/postnav :: NaviPost"></div>
     <div id="postContent" class="d-flex flex-column">
-        <div id="pageInfo" class="d-flex flex-column align-items-center align-self-center pb-4">
+        <div id="pageInfo" class="d-flex flex-column align-items-center align-self-center pb-3">
             <h1 class="py-2" th:text="${session.categoryList.get(categoryIndex).getName()}"></h1>
             <span class="py-1" th:text="${session.categoryList.get(categoryIndex).getDescription()}"></span>
+        </div>
+        <div id="tableInfo" class="mb-3 d-flex flex-row justify-content-start align-items-center">
+            <label for="sortStandard" id="sortLabel" class="me-2"><b>정렬기준</b></label>
+            <div class="dropdown d-flex align-items-center">
+                <select id="sortStandard" class="form-select dropdown-toggle" data-toggle="dropdown">
+                    <option class="dropdown-item" id="sortId" value="id">글번호</option>
+                    <option class="dropdown-item" id="sortTitle" value="bbsTitle">제목</option>
+                    <option class="dropdown-item" id="sortWriter" value="bbsWriter">작성자</option>
+                    <option class="dropdown-item" id="sortDate" value="bbsDate">작성날짜</option>
+                    <option class="dropdown-item" id="sortViews" value="bbsViews">조회수</option>
+                </select>
+            </div>
         </div>
         <h2 th:unless="${!bbsList.isEmpty()}">아직 글이 없습니다.</h2>
         <table id="table" th:if="${!bbsList.isEmpty()}"
@@ -30,13 +42,13 @@
             </tr>
             </thead>
             <tbody>
-            <tr th:each="bbs : ${bbsList}"
+            <tr th:each="bbs : ${bbsList.getContent()}"
                 th:onclick="|location.href='@{/post/bbs/view(id=${bbs.id})}'|">
-                <td style="width:8%;" th:text="${bbs.getId()}"></td>
-                <td style="width:30%;"><a class="bbs-link link-dark" th:text="${bbs.getBbsTitle()}"></a></td>
-                <td style="width:20%;" th:text="${bbs.getBbsWriter()}"></td>
-                <td style="width:20%;" th:text="${bbs.getBbsDate()}"></td>
-                <td style="width:8%;"  th:text="${bbs.getBbsViews()}"></td>
+                <td style="width:8%;" th:text="${bbs.id}"></td>
+                <td style="width:30%;"><a class="bbs-link link-dark" th:text="${bbs.bbsTitle}"></a></td>
+                <td style="width:20%;" th:text="${bbs.bbsWriter}"></td>
+                <td style="width:20%;" th:text="${bbs.bbsDate}"></td>
+                <td style="width:8%;"  th:text="${bbs.bbsViews}"></td>
             </tr>
             </tbody>
         </table>
@@ -48,23 +60,32 @@
         <nav aria-label="Page navigation" th:if="${!bbsList.isEmpty()}">
             <ul id="classUl" class="pagination justify-content-center align-items-center">
                 <li class="page-item btn" th:classappend="${!bbsList.pageable.hasPrevious()} ? 'disabled'">
-                    <a class="page-link"
+                    <a class="page-link" th:if="${#request.getParameter('sort') != null}"
                        th:href="@{/post/bbs(category=${#request.getParameter('category')}, page=${bbsList.pageable.previousOrFirst().pageNumber}, sort=${#request.getParameter('sort')})}">
                         Previous
                     </a>
+                    <a class="page-link" th:unless="${#request.getParameter('sort') != null}"
+                       th:href="@{/post/bbs(category=${#request.getParameter('category')}, page=${bbsList.pageable.previousOrFirst().pageNumber})}">
+                        Previous
+                    </a>
                 </li>
-                <th:block
-                        th:each="page : ${#numbers.sequence(bbsList.pageable.first().pageNumber+1, bbsList.getTotalPages())}">
+                <th:block th:each="page : ${#numbers.sequence(bbsList.pageable.first().pageNumber+1, bbsList.getTotalPages())}">
                     <li class="page-item btn"
                         th:classappend="${page == bbsList.pageable.getPageNumber()+1} ? 'active'">
-                        <a class="page-link" th:text="${page}"
+                        <a class="page-link" th:text="${page}" th:if="${#request.getParameter('sort')} != null"
                            th:href="@{/post/bbs(category=${#request.getParameter('category')}, page=${page}-1, sort=${#request.getParameter('sort')})}"></a>
+                        <a class="page-link" th:text="${page}" th:unless="${#request.getParameter('sort')} != null"
+                           th:href="@{/post/bbs(category=${#request.getParameter('category')}, page=${page}-1)}"></a>
                     </li>
                 </th:block>
                 <li class="page-item btn"
                     th:classappend="${bbsList.pageable.next().pageNumber == bbsList.getTotalPages()} ? 'disabled'">
-                    <a class="page-link"
+                    <a class="page-link" th:if="${#request.getParameter('sort') != null}"
                        th:href="@{/post/bbs(category=${#request.getParameter('category')}, page=${bbsList.pageable.next().pageNumber}, sort=${#request.getParameter('sort')})}">
+                        Next
+                    </a>
+                    <a class="page-link" th:unless="${#request.getParameter('sort') != null}"
+                       th:href="@{/post/bbs(category=${#request.getParameter('category')}, page=${bbsList.pageable.next().pageNumber})}">
                         Next
                     </a>
                 </li>
@@ -82,10 +103,10 @@
             const categoryId = [[${categoryIndex}]] + 1;
             //]]>
             changeUrl(categoryId);
+            sortBbs(categoryId);
             updateViews();
         }
     </script>
-
 </th:block>
 </body>
 </html>
