+++ src/main/resources/templates/post/writeBbs.html
@@ -48,7 +48,7 @@
     </div>
 </div>
 <th:block layout:fragment="script">
-    <script src="/js/post.js"></script>
+    <script src="/js/editpost.js"></script>
     <script th:inline="javascript">
         //<[CDATA[
         const categoryId = [[${categoryId}]];
